+++ jest.config.ts
@@ -1,0 +1,9 @@
+import type {Config} from '@jest/types';
+// Sync object
+const config: Config.InitialOptions = {
+  verbose: true,
+  transform: {
+  '^.+\\.tsx?$': 'ts-jest',
+  },
+};
+export default config;
