+++ package.json
@@ -25,6 +25,7 @@
 	"devDependencies": {
 		"@types/cors": "^2.8.12",
 		"@types/express": "^4.17.13",
+		"@types/jest": "^28.1.4",
 		"@types/node": "^18.0.3",
 		"concurrently": "^7.2.2",
 		"eslint": "^8.15.0",
@@ -36,6 +37,7 @@
 		"jest": "^28.1.0",
 		"nodemon": "^2.0.16",
 		"prettier": "^2.6.2",
+		"ts-jest": "^28.0.5",
 		"ts-node": "^10.8.2",
 		"typescript": "^4.7.4"
 	}
